refactor: move createSearchIndex to utils
